added update plane

+++ TransponderLib/Plane.cs
@@ -7,10 +7,9 @@ namespace ATM
         public string Tag { get; set; }
         public int XCoord { get; set; }
         public int YCoord { get; set; }
-        public uint Altitude { get; set; }
-        public uint Speed { get; set; }
-        public uint Course { get; set; }
-
+        public int Altitude { get; set; }
+        public double Speed { get; set; }
+        public double Course { get; set; }
         public DateTime LastUpdated { get; set; }
 
         public Plane()
